Fixing issues related to linking files while importing.
If we are just extrating a single file, do not try and rename it.  We could put this responsibility on the client, but this is more straightforward.  In the case where we choose to link a file, we still need to reset the zip file pointer to the correct location so that we correctly iterate through all the files in the zip.