Restore value of createNewFolder after unzipping single file.
The createNewFolder flag determines if the unzip process is responsible for creating, and possibly deleting the folder during the unzip process.  This is set to NO when unzipping into a library.  A call to unzipOneFile: was added during the general unzipping process, causing this value to always be set to YES, which meant that cancelling an import into a library would delete the entire library.

It is set to YES in unzipOneFile: so that it cleans up any files it makes in case something goes wrong.